discrete_control_requirements.txt
# ── JAX ecosystem ─────────────────────────────────────────────────────────────
jax[cuda12]>=0.9.1
flax>=0.12.1
optax>=0.2.6

# ── Atari environment stack ───────────────────────────────────────────────────
ale-py
gymnasium                   # keep only if your own code needs it
#dopamine-rl


# ── Computer Vision ───────────────────────────────────────────────────────────
# Used by dopamine atari_lib for frame preprocessing (resize, grayscale).
opencv-python>=4.10.0

# ── Configuration & logging ───────────────────────────────────────────────────
gin-config>=0.5.0
absl-py>=2.1.0

# ── TensorFlow ────────────────────────────────────────────────────────────────
tensorflow>=2.21.0

# ── Numerics ──────────────────────────────────────────────────────────────────
numpy>=2.0.0

# ── Experiment tracking (OPTIONAL) ────────────────────────────────────────────
wandb>=0.19.0
